+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 * [FIX] Handle dependencies with alternatives during build install
 * [FIX] config plugin now issues a warning when a given path is missing
 * [FEATURE] `before_build`
+* [FIX] config plugin now ignores symlinks #9
+* [CHANGE] service plugin now does not mark /etc/init.d/... symlinks as config #9